Added a watcher on the items property
I had an issue where I was loading a list of items from
an ajax request, and the directive was not set to take any action
after the items array changes. If the directive initialzed before
my ajax call finished, the directive would not get my items list
as it would not make it onto the scope until after the directive
had finished initializing.

Another use case I plan to explore more fully with this watcher
is the ability to apply custom property based filters to my
list in the controller scope, and then reassign the items property
to the directive and have it re-initialize with the new filtered
list.