Fixed issue with scroll boxHeight
Previously, if you filtered your list to less than the number
of items per page, the scroll box height might be set to be
larger than the scroll content area. This was due to
the math used to generate the box height. In most cases,
the number of filtered items divided by the itemsInPage would
be one or greater. In cases where the filtered items length was
less than the number of items in the page, a case could happen
where the boxHeight would be set to the scrollbar height divided
by a number less than one. This would cause the scroll box to
extend outside the area of the scrollbar.